chore: move tooltip ui component (#841)
