+++ packages/ui/src/components/TooltipRx/types.ts
@@ -1,0 +1,10 @@
+// Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+import type { AnyData } from '@polkadot-live/types/misc';
+
+export interface RadixTooltipProps {
+  children: React.ReactNode;
+  text: string;
+  theme: AnyData;
+}
